Install Oh My Zsh as an external archive
Leverages the .chezmoiexternal.$FORMAT functionality to automatically
install Oh My Zsh when running `chezmoi apply`.

The installation is based on grabbing the latest source as a tarball
and extracting the contents into the expected .oh-my-zsh directory.

Currently, the downloaded tarball is cached for 7 days (168 hours). This
prevents every execution of `chezmoi apply` from re-downloading the
tarball, which would have been detrimental to chezmoi performance.