Move $HOME/.local/bin PATH mangling to the end
As long as possible, we want user installed binaries to be towards the
front of the PATH environment variable. The easiest way to increase the
likelihood of this is to move the `export` statement as far down the rc
file as possible.

NOTE: We might want to consider moving it even further down. Perhaps at
the very bottom of the file.