Use LC_ALL=C when sorting printenv output
Using LC_ALL=C ensures that we get consistent and more intuitive sorting
behavior across different platforms and machines, regardless of locale.

In theory, setting LC_COLLATE=C should be sufficient, but it did not
work as expected on the machines I tested on. Therefore, we instead use
the nuclear option that is LC_ALL.