Use environment-specific EDITOR and GIT_EDITOR
This patch overrides the EDITOR and GIT_EDITOR environment variables
based on the terminal environment, e.g., VS Code integrated terminal or
SSH terminal session.

When in a VS Code integrated terminal we set `--reuse-window` to make
sure we open a new tab in the active window. Without this change most
git commands would hang around forever waiting for the window to close.

In an SSH session using a GUI editor does not make sense, so this commit
changes EDITOR and GIT_EDITOR to use nano instead.