Add silly hack to fix issue with direnv in VS Code
This patch adds a silly hack to resolve an issue with direnv
environments not being properly initialized when opening new integrated
terminal sessions.

Because integrated terminal sessions are launched from the workspace
root directory, direnv does not have an opportunity to detect when we
land in a direnv enabled workspace. The current workaround is to first
detect when a terminal has been launched in a folder that contains an
.envrc file. When that is true, we navigate to the parent directory and
then back to the original directory.

Hopefully we can find a nicer solution in the not too distant future...