Move a bunch of environment setup to .zshenv
This moves initialization of PATH and other such environment variables
from .zshrc to .zshenv. Probably not worth the move, but this makes
things a little bit cleaner perhaps. If nothing else, this should allow
me to override some variables based on whether the shell is interactive
or non-interactive, e.g., EDITOR.